read symlink and overwrite /usr/local/bin

+++ scripts/init.sh
@@ -33,7 +33,13 @@ mv ${TMPDIR} ${INSTALL_DIR}/ #copy setup script up to current dir
 # Symlink our scripts to /usr/local/bin
 cd /usr/local/bin/
 for file in ${INSTALL_DIR}/scripts/*; do
-  ln -s ${file}
+  name=`basename ${file}`
+  if [ `readink -e ${name}` != ${file} ]; then
+      rm ${name}
+      ln -s ${file}
+  else echo "Skipping file $name since symlink already exists"
+  fi
+
 done
 
 exec ${INSTALL_DIR}/scripts/$SETUP_SCRIPT
